test: configure Jest global setup and teardown
