+++ content/jest.config.unit.js
@@ -1,0 +1,8 @@
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+module.exports = {
+  preset: 'ts-jest',
+  silent: true,
+  testEnvironment: 'node',
+  testMatch: ["**/*.spec.(ts)"],
+  verbose: true,
+};
