+++ content/jest.globalTeardown.ts
@@ -1,0 +1,3 @@
+export default async () => {
+    await global.__POSTGRES_CONTAINER__?.stop()
+}
