+++ content/package.json
@@ -10,8 +10,8 @@
     "start:server": "node ./dist/src/entrypoints/run-server.js",
     "start:db": "node ./dist/src/entrypoints/run-local-database.js",
     "test": "run-p 'test:*'",
-    "test:unit": "jest ./test/unit",
-    "test:integration": "jest ./test/integration --runInBand --detectOpenHandles --forceExit --testTimeout=60000"
+    "test:unit": "jest ./test/unit --config=./jest.config.unit.js",
+    "test:integration": "jest ./test/integration --config=./jest.config.integration.js --runInBand --detectOpenHandles --forceExit"
   },
   "dependencies": {
     "@catalyst/commons": "0.1.0",
